use size_t instead of int for malloc and pointer calculations in iris_serde.c

+++ components/cloud/cloud.c
@@ -44,7 +44,7 @@ static esp_err_t mqtt_event_handler(esp_mqtt_event_handle_t event)
             iris_record_command(&ir_command);
 
             if (ir_command.length > 0) {
-				int data_size = 0;
+				size_t data_size = 0;
 				void * serialized_data = NULL;
 				serialize_data_from_ir_command(&ir_command, &serialized_data, &data_size);
 				free(ir_command.signal_pairs);
